Fix for the gui with gcc version 4.6.3.
There were casts in the gui from pointer to Int_t so as to save the pointer in
IDs. This was changed now so that the pointers are stored in a std::vector
while the ID is now the position at which the pointer can be found in the
vector.