First try at producing a python library.
Is not at all functional, but shows the direction one has to take. Probably
everything done in this commit should be reverted and then there should be a
separate directory for all the files similar to "python_stuff.cc", which are
then built with their own cmake file into their own library.