Starting from scratch.
Rebooted the whole business. Now working mainly in the isobarDecayTopology to
find all the permutations of final state particles which have to be considered,
their Clebsch-Gordans, etc.. Completely in the flux, just a commit to keep some
debug code somewhere.

Added a function to particleProperties, which allows to set the isospin
Projection instead of the charge.