A little work on the python bindings.
First of all changed to use python2.7, as boost.python for python3.2 refrains
from giving useful error messages when exceptions are thrown.

Also generated a binding for the particle data table which can be imported and
works, at least partly. It has been generated with Py++ which has been found to
be quite comfortable. Furter investigations are needed of course to see how one
could proceed.