A little progress on the python bindings.
Essentially started from scratch and this time started at the bottom. Produced
python bindings for the particleProperties and then on top the ones for the
particleDataTable. Also found a way to do that in different files, although
this may not be solved correctly at the moment.

The whole thing seems to compile and a superficial test with the library and
the python interpreter did not reveal major problem. Several functions do not
work though and should probably be excluded. A more systematic test should be
the next step.