Further on the python bindings.
Went over the particleDataTable now: rewrote most of the stuff and threw away
some of the more ugly code from PyPlusPlus. Also introduced a python interface
for the std::pair< const std::string, rpwa::particleProperties >. Together with
a renaming of the iterator in particleDataTable, this now allows to loop over
the table in the usual python style.

Also fixed a missing default argument in the particleProperties binding.