On to more advanced python bindings.
Re-introduced the wrapper struct for the particle properties. The way it is
done, the inheritance and overloading of particleProperties::qnSummary()
should now work. Also introduced a converter from a python list to a
std::set<std::string>, so functions which take such a set as an argument can
now be called in python with lists of strings. Tested this with
particleProperties::hasDecay(), but did not implement anywhere else yet.