More python bindings.
The particleProperties class is now fully accessible from python, including all
operators. particleProperties::chargeFromName() could not be directly ported,
as it uses a reference to int as an argument, which is an immutable type in
python. The signature has been changed for the python binding to return a
tuple, so the same functionality is available.

The particleDataTable is now only missing one method:
particleDataTable::entriesMatching(), which uses std::vector<> and std::set in
its arguments, so that will require some work still.