rpwa::particle python bindings and root class converters.
Created the python bindings for the particle class. This necessitated for the
first time the integration of ROOT/PyROOT classes, which got a little bit
messy. However, in the end I found a solution which is probably acceptable:
there are templated functions to convert to and from ROOT types. The only
caveat is that every specialization of these template functions has to be
explicitly exported to python, otherwise the library won't load.

The rpwa::particle bindings are not complete yet. Also made a small wrapper
around the .so library for easier loading into python and a small test script
to check the interface of rpwa::particle in python.

There was a small bug in the bindings for the particleProperties::qnSummary(),
fixed now.