Harmonized the test scripts.
Now there is only one test script which will test all the classes. The printed
output has been harmonized and reduced so that it is manageable. So far, the
test functions are far from exhaustive and no errors does not signify at all
that everything is working. It should give a good basis to do more elaborate
tests, though.