More python bindings.
Added the bindings for the interaction vertex together with some tests for it.
Also fixed a bug in the particle binding, where the comparison operator was
missing.

There seems to be a problem with the debug flags: the derived classes' debug
flags cannot be accessed from python. I introduced some tests for that and a
flag to skip these tests so that the problem will not be forgotten.