Fixed the debug flags in python.
Found out that static functions (and therefore properties) cannot be
overloaded with boost.python. The "solution" is to give each debug flag a
unique name, so now they are called "debug[CLASSNAME]". Fixed the test script
accordingly and that now seems to work.

Also fixed the path to the particleDataTable.txt in the test script, which
works now independently from the location it is executed.