Some minor changes to the python bindings.
Combined the different stlContainer export functions into one function, fixed
some formatting and a small bug in the test script. Also a small change in the
particle binding, where I am not sure if it changes anything but now it is
consistent with the way it is done in all the other classes.