Added the python bindings for the isobarDecayTopology.
Also added the tests for it, although most of them are skipped as I do not
construct a correct topology. The bindings necessitated the move of
constuctTopology to protected in isobarDecayTopology. Also, one more
stlContainer is needed.

I found out that clear() was not implemented correctly in the decayTopology
binding, which I fixed now.