Added the amplitudes' python bindings.
For the isobarAmplitude, the isobarCanonicalAmplitude and the
isobarHelicityAmplitude. However, there are only tests for the functions which
are defined in the inheriting amplitudes, not for the functions in
isobarAmplitude. Most of the tests would probably have to be skipped anyhow
until a valid topology has been implemented in the tests.