Python bindings for the waveDescription.
No tests were written for the wave description, as the next step will be to
write calcAmplitudes in python.

Also added a function to the isobarHelicityAmplitude's bindings which was
overlooked before.

Now all the essential classes from the amplitude package should have python
bindings. One known issue at the moment is that objects which are instantiated
in C++ code may not work when used in the python. For the moment, we probably
have to live with that but we might stumble upon a solution at some point.