Started with calcAmplitudes.py
Far from finished, but the argument parsing is there and some basic checks if
the arguments are valid. In the course of this, I fixed some bugs in the
waveDescription's bindings and added a python equivalent of the various print
functions to pyRootPwa.py. Also added a new binding file where some of the
reportingUtils get their bindings.

A rather important discovery is that one can produce bindings for the Write
method which the waveDescription inherits from TObject without adding anything
else from ROOT to boost python. This allows to store the waveDescription in
TTrees.

Also fixed a typo in orderEvtFile.py, which did not affect functionality.