Restructured the pyRootPwa package.
It is now a proper python package with the libRootPwa.so binding library only
being one component of it. The skeleton structure of the package is kept in
pyInterface/package and will be "built" into the pyLib directory. The package
can be imported as before, with the exception of the print functions, which
were moved to the utils submodule.

When "building" the python package, cmake uses a short python script
(python_compiler.py) which I hacked together. It mirrors the directory
structure of pyInterface/Package to build/pyLib/pyRootPwa. Then, all the *.pyc
python byte-compiled files are produced and moved to build/pyLib/pyRootPwa.
This is done whenever make is called and the pyRootPwa directory is cleaned
(rm -rf *) every time, so be sure not to put anything valueable in there...

This part should probably be revisited some time.