More code in calcAmplitudes.
Mostly changes in the config file and the arguments and the finding of input
files. The code added in this commit will probably have to be organized into
some smaller blocks at some point, but it will probably be useful as a
reference in any case.

I created a new file in the utils subpackage which can convert an .evt file
into a tree. Is not at all final yet, but the principle structure is there.

Also started a very simple calcAmplitude function in a new subpackage of
pyRootPwa. Thereby, I stumbled upon the problem that one cannot use objects
created in C++ as expected in python (the code from the wrappers is not
available), so amplitude calculation cannot work until this is fixed.