pyInterface: calcAmplitudes.py functional
calcAmplitudes.py is now calculating amplitudes. The restrictions at the
moments are that the output file cannot be in root format and the input file
has to be in root format.

The output file format configuration entry now can also be "root", thereby
removing the need to specify the file extension or file format anywhere else
than in this entry.

There is a new function in the _parseUtils which finds all the keyfiles to be
used. This is outsourced code from calcAmplitudes.py.

The rest are cosmetics changes.