pyInterface: Added support for root output files.
The "root" option for outputFormat now works and calcAmplitudes.py will produce
a root file with a TTree of amplitudeTreeLeafs if this option is chosen.

As a new branch on a TTree cannot be created directly in python if one wants to
save pyRootPwa objects, there is now an helper function in the
pythonAdministrator which takes over this task. This also necessitated the
introduction of a new rootConverter for the TTree.