pyInterface: Added ability to read .evt files.
.evt input files are now converted to memory-residing trees on the fly. There
is an info message and a progress bar showing the status of the conversion.
Unfortunately, at the moment the conversion of one input file is done multiple
times: once for every keyfile. Fixing this will need some restructuring, will
be done soon.

Also added a pyRootPwaException, but it is only used in the conversion from
.evt file to root tree, as it was not convenient to return a bool there.