pyInterface: Added the ampIntegralMatrix class.
No test for it yet. Also found some forgotten operators for the
amplitudeTreeLeaf binding, which I added now.

There are two functions not implemented yet, both have a comment as to the
reason. One returns a boost::multi_array, for which a conversion to python has
to be implemented first, and the other returns a
std::vector<rpwa::waveDescription>, which is not available in python do to
waveDescriptions's lack of operator==(). There is a fix for this, but it seems
it is not in the boost release version yet.