pyInterface: Added multi-threading support.
A first version to use the python multiprocessing module. It is only a very
rough sketch and will have to undergo quite some cleanup and some better
management of the input trees (they are cached completely at the moment), but
it is running correctly and producing output files. Speed-up on a 2-CPU machine
seems to be of the order of 40%.