pyInterface: Further multi-threading.
Fixed several issues from the last commit: now the input tree does not have to
be loaded completely into memory any more, but caching is supported with the
slice operator. There is some speed loss due to this, but it is not very big.
Also improved several other things which were not elegant, most notably the
printing of the status.

Also fixed the problems when processing .evt files, which did not work at all
in the last revision.