pyInterface: clean-up of multi-threading code.
The reading of the input files is now much cleaner and better arranged. Works
for both .evt and .root input files. Also changed the number of entries which
are read from the tree to 750, as this seems to be quite a nice number on my
test system. The dependence of the speed-up on this has to be investigated more
systematically though.