pyInterface: Fixed a bug in the inFile.
When more than one item of the tree was retrieved, the infile was returning a
list of TClonesArrays which was filled with copies of the same two entries.
This is fixed now by Cloning the TClonesArray before appending it to the list
to be returned, but produces some speed loss. Will investigate if this can be
recovered somehow...