pyInterface: progressBar optimization.
Optimized progressBar.updated() which produced significant speed up. As this is
one of the most called methods of the whole calculation, an option to disable
the progressBar altogether was introduced as well.

While I was on it, I commented out the command line arguments to calcAmplitudes
which are not implemented yet.