pyInterface: Getting the data in AmplitudeCalculator.
After the caching was removed from the inFile, the data is now actually read in
the AmplitudeCalculator as well. The inFile is now only a container for the
tree, with little actual functionality left. It cannot be discarded though, for
in the case of a .evt file the converted tree has to be transmitted to the
different processes working with it.