pyInterface: Introducing data output caching.
Changed the AmplitudeCalculator.calcAmplitudes() function so that it caches the
calculated amplitudes in a python list. Appending values to this list is faster
than using the array.array. The number of events to write to this list before
writing to disk is specified in the config file. Also removed some old config
file stuff which was left over from the input file caching.

The python script now appears to be less than 5% slower than the
shell-script/c++ version (with the progressBar disabled, that is).