pyInterface: Profiler can be enabled by option.
The python profiler module can be enabled by command line option to python now.
It essentially disables all multiprocessing features and runs the functions
normally run in its own process directly. Also changed the optional arguments
to pyRootPwa.bin.calcAmplitudes() to be a dictionary, which makes it cleaner,
especially as more arguments are to be added.