pyInterface: Major restructuring.
The pyRootPwa package was restructured for easier and more intuitive use. The
following things were changed:

- pyRootPwa.bin.calcAmplitudes -> pyRootPwa.calcAmplitudes (deleted bin
  subfolder)

- The rootPwaConfig and rootPwaException are now available directly in the
  pyRootPwa scope

- The inputFile was moved to the amplitude submodule

- All python/c++ bindings were moved to pyRootPwa.core

- Silencer -> silencer, AmplitudeCalculator -> amplitudeCalculator

- Revised the __init__.py scripts, now only the actual pyRootPwa objects should
  be imported and visible in python (before, e.g. sys was available in various
  submodules of pyRootPwa)

- Some file renaming for more consistency

There might still be some changes with regards to structure (e.g. move some
functions from the utils to the amplitude submodule), but the idea would be to
not touch the general structure any more. If more functionality is added, this
should lead to the addition of submodules and objects in the pyRootPwa scope
and should not lead to existing submodules changing.