correctly take tprime limits into account
as previously already done for the tMin variable, also the tprimeMin and
tprimeMax variables are now always stored with a negative sign. This is
enforced in the setter, if the provided value is positive.
Also the variables t and _tprime are using this convention, the allowed
range of tprime is now correctly and from both sides (upper and lower
limit) checked.