Introduce new class generatorManager.
Add a new class called generatorManager, which should take over most of the
work now done in the main of genpw. Began writing the code to parse the
reaction file, not completed yet but a good portion of it is there.

This commit breaks genpw! The reaction file is not read any more as I am using
genpw now to test the new code in the generatorManager. At some point, genpw
will be cleaned out and then of course become functional again, but there will
be a few commits now where it will not work.