Move helper classes to their own file.
Move three helper classes, which were defined as subclasses of the
generatorManager until now, to their own files. Like this, they can be used
later as arguments when the generator parameters are set.

Also make the beam section in the reaction file mandatory. While the beam
simulation package is optional, the beam parameters are also needed if the
package is not used. As everything is mandatory now, change from pointers to
"real" objects to save these information, which makes handling somewhat easier.