Add a generator class and use parameter classes.
Introduce a new generator class which is used to define the interface an event
generator should use. The diffractivePhaseSpace is now inheriting from this
class. Also change over to use the new parameter classes to get information
into the generators, this is done already on the level of the base class.

There is quite some commented code around now, will clean up at a later point.