t'- and m-distribution parametrization.
Add the possibility to parameterize the t'- and m-distributions of the final
state. One can specify in the reaction file which function should be used to
produce the desired distributions. The functions are defined in
generatorPickerFunctions.hpp, a common interface they have to inherit from is
also there. The generatorManager instantiates the appropriate class (as
specified in the reaction file) and passes the class-specific functions to it.

To keep functionality which is already there, implement a parametrization to
have a uniform mass distribution with a exponential t' distribution. The slope
of the t' distribution can be specified in mass slices. The values given will
then be interpolated linearly.