Improve config parsing for the pickerFunction.
For the uniform mass - exponential t' picker function, rename the t_min
parameter to t_prime_min and introduce a new t_prime_max parameter. Make both
of them optional with the appropriate default values (0 and infinity). Also
introduce a safeguard so that the function is not used without having been
initialized before.