Fix a problem with the random number generator.
When the random number generator was removed in the nBodyPhaseSpaceGen, the
member function "random()" was also removed, causing parts of the code to fall
back to std::random(). This problem was a good motivation to rename the
"random()" function of the randomNumberGenerator, sot the problem will not
occur there as well. Make the nBodyPhaseSpaceGen use the correct random number
generator.