Write events to ascii functions and use particle.
Move the functions which write the ascii event files from the
diffractivePhaseSpace to the generator. Implement them as static functions, so
convert the generator.hpp to a header and source file. As the particle class
holds a TLorentzVector (which the particleProperties class does not), switch to
the particle class for initialization as well as storing the generated data.