More interface for the generatorManager.
Prepare the interface of the generatorManager class for the loop to produce
events. Add a function to generate an event and to get a reference to the
generator to extract the generated information. Also introduce the possibility
to override the mass range set in the reaction file, which is useful if one
wants to generate events in mass bins.