Add warnings and info about removed features.
In genpw.cc, add warnings about command line options which use features
which have been (temporarily) removed: add a warning that the features
are gone and information on how to get them back in the help text and
print an error message if such a option is invoked. In the
generatorManager, issue warnings if options in the reaction file are
found which were supported before and are not anymore.