Generate events in genpw.
Implement a loop to generate events in genpw. It can now write the ASCII and
Comgeant event files. There are hundreds of lines of more or less commented
code from the old version still in, clean-up is pending.

Also make t' negative in the diffractivePhaseSpace. I am not entirely clear why
this is necessary, but it seems to do the right things now.