Add a new mass and t' picker.
With the new picker, one can give coefficients of two polynomials in the
reaction file. One of the polynomials determines how the masses are
distributed. The other polynomial returns a t' slope for the mass given
as argument. With this, it should be possible to approximately model the
t'/mass distribution.

There is quite some duplicate code (reading the mass- and t'-ranges from the
reaction file), which should be reorganized.