pyInterface: Completely revise all exposed classes.
There were still some unresolved problems in the bindings for python: objects
instantiated in C++ parts of the code were not behaving correctly when used in
python. The problems originates in the fact that the wrapper classes were also
used to solve problems with translation of arguments between python, C++ and
PyROOT. These translations were then not available for the python part of the
code, because the objects from C++ were not wrapper classes.

Solve this problem now: use the wrapper classes only for inheritance issues,
which fixes all the functions which need neither translation work nor are
virtual. For the functions which are virtual, introduce an additional function
definition, pointing to the C++ class, which will take effect if the object was
instantiated in C++. For the functions which need translation work, use a
global function which takes a reference to the class as first argument (the
python-style "self"). If both inheritance and translation work is needed, do
the translation in the wrapper as before, but also introduce the global
function, which is used for objects from C++ (unfortunately, this is some code
duplication, but I do not see a possibility to avoid it).

Follow this approach for all exposed classes. Remove the wrapper completely if
it is not needed.