pyInterface: Bugfix.
There was a bug in amplitudeCalculator: if the output format was root or ascii,
the real and imaginary parts of the amplitude where written to the real parts
of two consecutive events, with the imaginary parts always zero. Change the way
the amplitudes are cached to solve the problem. Unfortunately, this requires
some acrobatics to keep the binary output working as before.