pyInterface: Add tests for waveDescription.
Add a number of tests for the waveDescription. Also change the interface for
waveDescription::constructDecayTopology to match the one of
constructAmplitude() (return a tuple with the result and the constructed
topology).

The waveDescription tests enables the use of a test keyfile from which valid
amplitudes and topologies are constructed, which can then be used for other
tests (which are currently skipped).