amplitude: Different comparison for some maps.
For maps where a boost::shared_ptr is used as key, there is a problem with
boost::python: a pointer which has made a trip through python cannot be used
anymore to retrieve objects from the map. This breaks decayTopology::subDecay.
To overcome the problem, introduce a comparator struct in the decayGraph, which
simplifies the comparison to a point where the pointers from python work again.