Better output from the angle plotter.
Change the command line interface to be similar to the one from
calcAmplitudes.py. The input data is now determined by a command line argument
which gives the bin number. The data from the bin range given will be combined
for one set of histograms. The option can be given multiple times, the
histograms are then stored in separate folders in the output file.