pyInterface: Improve tests.
Some tests involving properties could produce a false positive if the property
was not present (i.e. boost::python did not produce it). This happened because
some tests assigned something to the property before reading them. If the
property was not there in the first place, the tests created them themselves,
so the test would then succeed. Changed this for the particle's tests, but
still have to check if this could happen also in other places.

Also, make the tests abort if the module could not be imported. Neglecting to
do so will produce a segmentation violation (it should not, but I suspect that
to be a quirk of PyROOT).

Fix the particle::lzVec test, make sure isobarHelicityAmplitude::hfTransform()
is really working and add a test for isobarAmplitude::gjTransform().