pyInterface: Fix memory leak.
There was a nasty memory leak in the python interface: whenever a root object
was returned in python, the C++ object was copied. The copy was not deleted
when it went out of scope in python. If such a function was called in a loop,
python filled up memory very quickly.

To fix the bug, a new feature was needed in pyroot, which was added in root
version 5.34.04. Change the CMakeLists so that the python bindings are only
built if at least this version was found.