amplitude: Refactor amplitude symmetrization.
Reorganize the code which deals with the Bose- and isospin symmetrization of
the amplitude: move the code which finds the necessary permutations to the
isobarDecayTopology. This necessitates a change in the code which combines the
Bose- and isospin-symmetrizations.

There is a new function in the isobarDecayTopology which is not yet integrated
in the python interface. Will be added soon.