amplitude: Add functions to isobarDecayTopology.
Add two more functions to the isobarDecayTopology:

1. isobarDecayTopology::isobarIsAffectedByPermutation, which returns true if
the given vertex (i.e., the isobar decaying in it) will have changed kinematics
if the permutation is applied.

2. isobarDecayTopology::getFsPartIndicesConnectedToVertex, which returns the
indices of all the final state particles which are "below" the given vertex.
This is used as a helper function for 1., but make it public as it might prove
useful also in other situations.

Add the functions to the python interface and provide tests for them. They will
be used in the angle plotting script.