Generators: Fix t' slope interpolation.
The determination of the t' slope for a given invariant X mass was done in a
step function. Improve this by interpolating if the mass is given in between
two values from the reaction file. Masses below or above the range covered in
the reaction file will be assigned the first or last t' slope given,
respectively. Also adapt the printing to the interpolation so that it is
understandable what is going on.