Generators: Fix NaN in output files.
If the weight in the nBodyPhaseSpaceGen is calculated to be NaN, set the weight
to 0 (instead of just printing a warning and leaving the weight as it was).
This leads to the rejection of such events (before, they were written to the
output files).

Fix for bug #5.
(cherry picked from commit 621d9642940ad37f9406ee41bb6bbedb48859bb4)

Conflicts:
	generators/nBodyPhaseSpaceGen.cc