Utilities: Check for rounding errors.
Further investigation into the cause of the NaN values in the
nBodyPhaseSpaceGen (see bug #5 and commit 621d964) revealed that due to
problems with the machine epsilon, breakupMomentumSquared() in physUtils.hpp
could return (very small) negative values which turned to NaN when the square
root of them was taken later. Fix this by explicitly checking if the squared
brakeup momentum is smaller than zero (setting it to zero if it is) before
returning it.