Change printSvnVersion to printGitHash.
Change the CMakeLists.txt file so that it will no longer try to set the svn
version. Instead, get the latest git hash and add it as preprocessor variable.
Change the name of this variables from SVN_VERSION to GIT_HASH and change the
name of the printSvnVersion function to printGitHash.

Propagate this change to the python interface and add a test for the printing
functions which are exposed in python.