pyInterface: Add to/from vertex to python interface.
The toVertex() and fromVertex() functions are defined in the decayGraph, which
is not part of the python interface. However, as toVertex and fromVertex are
such useful functions, add them as member functions of the decayTopology to the
python interface. Also add tests for them.
(cherry picked from commit 57d5c5f6b563efc9540a85df5c1c47f788362f5e)