anglePlotter: Fix Bose symmetrization.
There was a fundamental problem in the way the Bose symmetrization was done in
plotAngles.py. Fix this now. As the correct way of doing it is quite some lines
longer than the previous one, move it to its own function.
(cherry picked from commit 80401d1c6a4919ac12233cbd544aa1ee50746189)