python Scripts: Fix shebangs.
Change the shebangs of all python scripts to "#!/usr/bin/env python". This will
cause the scripts to be executed with the same python interpreter that would
start up if one types "python" on the command line. This could be a problem if
rootpwa is compiled with another version of python than it is linked to be
executed when "python" is invoked, but it is better than what was used before.
This should probably be revisited in the future.

For the gitUtils/serverHook.py, the shebang reads "#!/usr/bin/env python2.6"
because it is supposed to run on the sourceforge server.