evtweight: weights for inidividual waves
storing the weight of each individual wave to the TTree drastically increases
the file size (for the 88 waves wave set by a factor of 7-8 compared to only
the events). only storing the (sum of) weight of the waves with positive or
negative reflectivity, and the flat wave only adds 30%. therefore the later now
is the default, while the first option can still be actived with the command
line option '-s'.