Generators: Delete beam simulation package.
Delete all the code in the primaryVertexGen class so that only a skeleton is
left. Rename the class to beamAndVertexGenerator and rename the variables
holding it accordingly. The files will be renamed in the next commit (so that
it will be a "clean" rename). And new beam simulation package will be added
over the course of the next few commits.