Generators: Add new beam simulation package.
Add the code for the new beam simulation package. It takes a root beamfile
with a tree of events and sigmas. To generate an event, an entry of the tree is
randomly chosen and the event it contains is smeared with the corresponding
sigmas. This way, one gets the vertex X- and Y- position and the beam momentum.
The vertex Z-position is determined by the length of the target and the
radiation length (i.e., there are less events generated towards the end of the
target if the radiation length is bigger than zero). For this, there is a new
mandatory option in the target section where the interaction length of the
target is given.

Also apply this method of determining the vertex Z distribution in the
generation without the new beam simulation package.

Change the behavior of the generatorManager if an error is encountered during
the initialization of the beam simulation package: before, a warning message
was issued and the generator proceeded with the simple beam simulation. Now,
issue an error message and abort.