pyInterface: Fix tests.
There were two tests failing because of a change in utilities/physUtils.hpp
introduced in commit bbf69d9. Due to this change, some of the amplitudes no
longer turn out to be NaN, which is what was tested for in testPyRootPwa.py.
Fix this by comparing the calculated amplitude against the correct reference
value.