pyInterface: Fix tests.
There were two tests failing because of a change in utilities/physUtils.hpp
introduced in commit bbf69d9. Due to this change, some of the amplitudes no
longer turn out to be NaN, which is what was tested for in testPyRootPwa.py.
Fix this by comparing the calculated amplitude against the correct reference
value.
(cherry picked from commit 5f4704f1fef9d0564c47077e9697978bfaba1ca9)

Edit: on this branch, the commit introducing the problem was 9cffd9c.