weighting of 3pin: preparing overlay of weighted PS and accepted PS MC
Added a second wrapper function that accepts a third filename, this should then
be called with the name of the file containing the real data, the weighted
phase-space events, and the weighted accepted phase-space events (in that
order). Reversed the order in which the trees are processed. First process the
real data tree, then the other(s), order of the histograms had to be changed
accordingly.