Generators: Fix a bug and some cosmetics.
Remove a double delete in the generator which led to a crash when using a
beamfile. Also make sure the the beamAndVertexGenerator::event() returns true
and that this return value is checked (it could be used in the future to catch
possible errors during the beam and vertex generation). Rename a variable in
the generation manager.