Create acceptance histograms
If running with both, weighted accepted MC events and weighted PS MC events,
then compute the ratio for each pair of histograms to get the acceptance. For
this some more things have been cleaned. A second list of histograms from which
the acceptance should be calculated was added, the processing of each of these
lists has been unified to process TH1 and TH2 in the same loop.
Also only store one cycle of the objects written to the file.