Generators: Move vertex Z picking function.
Move the function which picks out the Z position of the generated primary
vertex. It was implemented twice, once in the diffractivePhaseSpace and once in
the beamAndVertexGenerator. Move it now to the abstract generator class to
avoid code duplication. This has one caveat though: for the
beamAndVertexGenerator to have access to it, the generator needs to pass a
reference to itself when calling beamAndVertexGenerator::event().