Generators: Fix vertex Z generation.
Change the implementation of the generation of the Z position of the primary
vertices. Before, there was linear decrease of the number of vertices over the
target, which was wrong. Implement an exponential decrease of the number of the
vertices, with the decay constant being the interaction length, which is the
right approach.