utils: Check if libconfig source path exists.
In the libConfigUtils, when the path to the source file was retrieved for use
in error messages, it was not checked if it was a NULL pointer (return type was
const char*). This could lead to silent crashes when the Setting object was not
read from a file. Fix this by introducing a new util function which gets the
source file path, checks for NULL and returns a std::string (which will be
empty if the pointer was NULL).

Fix for bug #17.
(cherry picked from commit 4699d87dd0f20486ef0ab1affd205d94f696d5a2)