Fix evidence calculation in massbins with thresholds
The calculation of the evidence for a massbin with waves below threshold
returned NaN, the reason seem to be rows and columns in the covariance matrix
containing nothing but 0s, and a call to ROOT::Math::normal_cdf_c with a sigma
of 0.

Identify the thresholded waves by testing the real and imaginary part of the
corresponding production amplitude to be zero. Do not consider thresholded
waves in the covariance matrix, the parameter volume, and while checking for
negligible waves.

Signed-off-by: Boris Grube <bgrube@tum.de>