Generators: More output for the beam package.
Add more output to the beam and vertex package, it is now also printed if the
beamfile is read sequentially. Unfortunately, this always prints "No" because
the option is applied only after the printout has happened. As this is very
confusing, I will generate a bug report on that and it will be fixed at some
point.