Generators: Introduce better printing.
Introduce a new print function to the generatorManager class, which prints all
the information on all the parameter structs and classes it knows. Remove the
same printouts from readReactionFile(). Use the new print function directly in
genpw after all the initializations have been made. This solves the problem of
the confusing printouts because the printouts happened before the overrides
took place.

Fix for bug #21.