Generators: Replace asserts and add an option.
Add an option to genpw with which the sequential beamfile reading can be
started from a random position in the beamfile. This makes it possible to have
random beamfile events without sacrificing computing time.

Also replace several assert()s with if()s so that the corresponding checks are
executed even with NDEBUG defined. Thanks to Sebastian Uhl for pointing these
out to me.