Generators: Fix problem in beam package.
When the vertices were generated, the determination of the x- and y-positions
was done completely independent of the z position and the momentum. Now, assume
that the x- and y- positions in the beamfile were projected to z = 0 and then
extrapolate to the generated z position using the x- and y-components of the
momentum.