Added a picker for tPrime distributions with multiple exponantials
A new picker 'uniformMassMultiExponentialT' has been introduced, it allows to
generate events according to a flat mass distribution, and a tPrime
distribution that may consist of several exponantials:
exp(p[0]*t) + sum(p[2i-1]*exp(p[2i]*t), i=0...n)
The parameters are read from the reaction file, the 'invariantMasses' is as
for the 'uniformMassExponentialT' picker, the 'tSlopes' option is a list of
either numbers, or arrays of numbers.

As using ROOT's TF1 to get the random tPrime turned out to be rather slow, a
numeric method has been implemented here.