Generators: Test parameters and improve getting tPrime in a picker.
Test the parameters used to calculate the tPrime slope after reading from the
reaction file. Initialization is stopped if a wrong parameter is found. In case
the a random tPrime cannot be obtained in 50 steps of the Newton-Raphson
method, print a warning and restart with a new random value. Do this five times
before finally giving up. Also fixed two closing quotation marks in error
messages.