Generators: Increase efficiency.
Increase the efficiency of the diffractive phase space generator by calculating
several maximum weights for mass slices and then setting the max weight later
as appropriate for the X mass which was picked by the picker functions.
Depending on the mass range which the generator has to cover, this can give
huge increases in efficiency.