amplitude: added a flatRange mass dependence
A new mass dependence 'flatRange' was added. This amplitude is mostly 0
unless in a range of the width of the mother state around the mass of the
mother state where it is 1. Such an amplitude is required, e.g., to study the
pipi S-wave in the 3 pi final state, or similar.