Added 2D angle plots in Gottfried-Jackson and helicity frame
Added 2D plots comparing cos(theta) and phi in the Gottfried-Jackson and the
helicity frame between real-data and weighted MC. With recent changes to the
evtweight program rescaling of histograms is no longer required. Labeling of
the z-axis of 2D histograms has been improved, also for 2D ratios, the label
of the z-axis is modified, not the one of the y-axis. Also some cleaning was
done.