amplitude: New function for particleProperties.
Add a new function to the particleProperties class: isStable(). In the current
implementation, it will return true if the particle in question is a pion and
false for all other particles. This is done as preparation for the
implementation of the amplitude parametrization for the case where the daughter
particles are unstable.